+++ tests/test_deprecations.py
@@ -8,29 +8,30 @@ from baybe import BayBE, Campaign
 from baybe.searchspace import SearchSpace
 from baybe.strategies import Strategy
 from baybe.targets import Objective
+from baybe.targets.base import Target
 from baybe.utils.interval import Interval
 
 
 def test_deprecated_baybe_class(parameters, objective):
-    """Using the deprecated ``BayBE`` class should raise a warning."""
+    """Using the deprecated ``BayBE`` class raises a warning."""
     with pytest.warns(DeprecationWarning):
         BayBE(SearchSpace.from_product(parameters), objective)
 
 
 def test_moved_objective(targets):
-    """Importing ``Objective`` from ``baybe.targets`` should raise a warning."""
+    """Importing ``Objective`` from ``baybe.targets`` raises a warning."""
     with pytest.warns(DeprecationWarning):
         Objective(mode="SINGLE", targets=targets)
 
 
 def test_renamed_surrogate():
-    """Importing from ``baybe.surrogate`` should raise a warning."""
+    """Importing from ``baybe.surrogate`` raises a warning."""
     with pytest.warns(DeprecationWarning):
         from baybe.surrogate import GaussianProcessSurrogate  # noqa: F401
 
 
 def test_missing_strategy_type(config):
-    """Specifying a strategy without a corresponding type should trigger a warning."""
+    """Specifying a strategy without a corresponding type raises a warning."""
     dict_ = json.loads(config)
     dict_["strategy"].pop("type")
     config_without_strategy_type = json.dumps(dict_)
@@ -39,12 +40,25 @@ def test_missing_strategy_type(config):
 
 
 def test_deprecated_strategy_class():
-    """Using the deprecated ``Strategy`` class should raise a warning."""
+    """Using the deprecated ``Strategy`` class raises a warning."""
     with pytest.warns(DeprecationWarning):
         Strategy()
 
 
 def test_deprecated_interval_is_finite():
-    """Using the deprecated ``Interval.is_finite`` property should raise a warning."""
+    """Using the deprecated ``Interval.is_finite`` property raises a warning."""
     with pytest.warns(DeprecationWarning):
         Interval(0, 1).is_finite
+
+
+def test_missing_target_type():
+    """Specifying a target without a corresponding type raises a warning."""
+    with pytest.warns(DeprecationWarning):
+        Target.from_json(
+            json.dumps(
+                {
+                    "name": "missing_type",
+                    "mode": "MAX",
+                }
+            )
+        )